requirements.txt
# Base requirements
instagrapi>=2.1.3
python-dotenv>=1.0.0
requests>=2.32.3      

# Add Celery and Redis client
celery>=5.0.0
redis>=3.5.0